+++ src/main/java/pers/goetboy/entity/sys/RoleMenu.java
@@ -25,12 +25,12 @@ public class RoleMenu extends AbstractEntity implements Serializable {
      * 角色id
      */
     @Column(name = "role_id")
-    private Long roleId;
+    private Integer roleId;
     /**
      * 菜单id
      */
     @Column(name = "menu_id")
-    private Long menuId;
+    private Integer menuId;
 
 
 }
